refactor: split ExcelSheet into partial files
